--- src/components/shared/FormUncontrolled/form.uncontrolled.module.scss
@@ -1,34 +1,0 @@
-.title {
-  margin-bottom: 4rem;
-}
-
-.submit {
-  width: 100%;
-}
-
-.checkbox {
-  font-size: 1.4rem;
-}
-
-.root {
-  display: flex;
-  flex-direction: column;
-  gap: 0.8rem;
-  width: 100%;
-}
-
-.label {
-  font-size: 1.4rem;
-  color: var(--text-secondary);
-  margin-left: 0.4rem;
-}
-
-.selectContainer {
-  position: relative;
-  width: 100%;
-}
-
-.passwordStrength {
-  margin-top: 2.4rem;
-  padding-left: 1.5rem;
-}
