+++ src/styles/variables.scss
@@ -10,7 +10,7 @@ $ff-primary: 'Poppins', sans-serif;
 
   // Functional Colors
   --color-success: #00c853;
-  --color-warning: #ffd600;
+  --color-warning: #d8b401;
   --color-error: #ff3d71;
   --color-error-bg: rgba(255, 61, 113, 0.1);
   --color-success-bg: rgba(0, 200, 83, 0.1);
