+++ src/types/country.ts
@@ -1,0 +1,4 @@
+export interface Country {
+  code: string;
+  name: string;
+}
